\name{inc.mapkey}

\docType{data}

\alias{inc.mapkey}

\title{Acronyms, HEX colours and shapes recommended for mapping outbreak data}

\description{
Acronyms, HEX colours and shapes recommended for mapping outbreak data.}

\usage{data(inc.mapkey)}

\format{
  A data frame comprised of 10 rows with the following four variables.
  \describe{
    \item{scode}{the two or three letter code to define premises status.}
    \item{sname}{a descriptor of premises status.}
    \item{shape}{the shape to be used for plotting premises status.}
    \item{hex}{the HEX code to be used for plotting premises status.}
  }
}

\references{
Animal Health Australia, 2022. Disease Strategy: Foot and mouth disease (Version 5.0). Australian Veterinary Emergency Plan (AUSVETPLAN). Primary Industries Ministerial Council Canberra, ACT.
}

\seealso{
The documentation for data set \code{\link{inc.aucsf}} provides code to produce a situation report map using the recommended map symbology.
}

\keyword{datasets}
