\name{inc.about}

\alias{inc.about}

\title{
The library incursion: summary information
}

\description{
Functions for the analysis of infectious disease outbreaks in human and animal populations.
}

\usage{
inc.about()
}

\details{
The most recent version of the incursion package can be obtained from: \url{https://mvs.unimelb.edu.au/research/groups/veterinary-epidemiology-melbourne}.
}

\author{
Mark Stevenson (\email{mark.stevenson1@unimelb.edu.au}), Melbourne Veterinary School, Faculty of Science, University of Melbourne, Parkville, Victoria 3010, Australia.

Robert Sanson, AsureQuality New Zealand, PO Box 1654, Palmerston North 4410, New Zealand.

Andrew McFadden, Investigation and Diagnostic Centre Biosecurity New Zealand, PO Box 40742, Upper Hutt 5018, New Zealand.

Matthew Edwards, EpiVets New Zealand, 565 Mahoe Street, Te Awamutu 3800, New Zealand.
}
